Moved installation of choco package to separated helper
